Draft: Split DraftLInk from Draft.py
